Create Isomorphic Strings
Two strings s and t are isomorphic if the characters in one string  can be replaced to get another string.
All occurrences of a character must be replaced with another character while preserving the order of characters. No two characters may map to the same character, but a character may map to itself.